python/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lz==0.11.1
MarkupSafe==2.1.3
memoir==0.0.3
mpmath==1.3.0
multidict==6.0.4
mypy==0.991
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.1
openai==0.27.8
packaging==23.1
paradigm==0.6.2
Pillow==10.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
reprit==0.9.0
requests==2.31.0
safetensors==0.3.1
six==1.16.0
son==0.4.2
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.31.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
yarl==1.9.2
